get the file size

+++ gd_cache.h
@@ -34,7 +34,7 @@ struct gd_fs_entry_t {
 	char *filename; // 'title' in the XML from directory-list
 	char *resourceID;
 
-	size_t size; // file size in bytes, 'gd:quotaBytesUsed' in XML
+	unsigned long size; // file size in bytes, 'gd:quotaBytesUsed' in XML
 	char *md5; // 'docs:md5Checksum' in XML, might be useful later
